refactor: extract RecordingSessionPresenters.swift (#650 slice A)
Second slice of #650. Extracts the 4 presenter classes
(RecordingSessionStartStatusPresenter, RecordingSessionCancelStatusPresenter,
RecordingSessionStopReadinessPresenter, RecordingSessionStopFailurePresenter)
from RecordingSessionController.swift into a new
RecordingSessionPresenters.swift.

Byte-identical move; SHA-256 1ca2b0e5bb0919232176bb64feeaa6a6f6ff2b805678f781e62563700beb44fd
verified against origin/main:RecordingSessionController.swift lines 3-189.

Reduces RecordingSessionController.swift from 406 to 219 lines.

Tests: 667 (unchanged).
